feat: create two status section component
